+++ docs/changelog.rst
@@ -18,6 +18,21 @@ To avail of fixes in an unreleased version, please download a ZIP file
 `directly from GitHub <https://github.com/mitogen-hq/mitogen/>`_.
 
 
+In progress
+-----------
+
+* :gh:issue:`1539` :mod:`ansible_mitogen`: Fix "module 'dnf' has no attribute
+  'cli'" on Ansible 14
+* :gh:issue:`1540` :mod:`ansible_mitogen`: Remove override of
+  ``ansible/__init__.py``
+* :gh:issue:`1540` :mod:`ansible_mitogen`: Remove ``reload(sys)`` workaround
+* :gh:issue:`1540` :mod:`ansible_mitogen`: Refactor "fixups" as source modifiers
+* :gh:issue:`1540` :mod:`mitogen`: Add :class:`mitogen.master.ModuleFinder`
+  source modifier feature
+* :gh:issue:`1540` tests: Test :class:`mitogen.master.ModuleFinder` source
+  override
+
+
 v0.3.51 (2026-07-18)
 --------------------